feat: Archive 페이지 공개 버전 구현
